End of day 18 dec. Needs a close look to waveType selection.

+++ CSD2b/les4/synthesizer/sineWave.cpp
@@ -14,7 +14,7 @@ SineWave::~SineWave() {
 
 //Functions
 
-int SineWave::getSample() {
+float SineWave::getSample() {
   std::cout << "SineWave || getSample()" << std::endl;
   return sin(phase);
 }//getSample()
